chore: Move extslices to utils and rename to slicesutils
