+++ internal/utils/slices/slices.go
@@ -1,4 +1,4 @@
-package extslices
+package slices
 
 func TransformFunc[S1 ~[]E1, E1, E2 any](s S1, transform func(E1) E2) []E2 {
 	results := make([]E2, len(s))
